requirement.txt
asgiref==3.5.2
Django==3.2
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
termcolor==1.1.0
typing_extensions==4.3.0
zipp==3.8.1
